+++ CHANGELOG.md
@@ -123,6 +123,8 @@ All notable changes to this project will be documented in this file.
 -   #2175 : Fix the shape of multi-level containers.
 -   Catch all internal errors arising in the syntactic, semantic, printing or code generation stages.
 -   #2206 : Fix returning an array of unknown literal size.
+-   #2112 : Improve floor division.
+-   #2214 : Fix returning a local variable from an inline function.
 -   #1321 : Fix use of tuples returned from functions in a non-assign statement.
 
 ### Changed